asfdemux: Fix prerolling files with "empty" streams
This is a regression that was introduced by
commit 1803b3c18530cb0100d140c2b8e49a8dfe41f941
"    asfdemux: Add support for dvr-ms"

The problem is that some files/streams might contain stream definition
but there is no actual packets for those streams.

This was used to "define" streams with different bitrates for example.

The first_ts calculation resulted in never ever finding a valid first_ts
since some streams were empty, and therefore never "activating" itself.

Instead of that we first check if we are prerolled. And if we are we
unconditionally get the "first_ts"

The preroll check has been adapted to check whether streams of
each defined type (i.e. audio/video/sub) has been prerolled. This solves
the problem of having different streams of a particular type where only
one stream actually has data.